gensrc: extend search for call instruction in warnings detection code
Currently warnings detection code looks for a call to warn_slowpath* functions
only in the following 4 lines after line-number saving instruction.  But there
are warnings with longer distance (5-6 lines) in some functions that are better
left unpathced (e.g. tcp_recvmsg).  This change increases searched distance by
two lines with some detection code reworking.

Skipping of checked lines is also fixed.  Currently, after successfull detection
of the warning code, iterators are set to the line number following the "call
warn_slowpath*", which is wrong since in for-loop in cblock_cmp() these iterators
are incremented in the adjustment expression. So it's possible that the lines that
follow "call warn_slowpath*" will not be compared.

Change-Id: I57087212762284c9a5097b6a8a8a87d4886914b2
Jira-ID: https://cloudlinux.atlassian.net/browse/KCARE-426
Signed-off-by: Volodymyr Lukiianyk <vlukiianyk@cloudlinux.com>
Reviewed-on: https://gerrit.cloudlinux.com/20170
Reviewed-by: konstantin olshanov <kolshanov@cloudlinux.com>
Tested-by: konstantin olshanov <kolshanov@cloudlinux.com>