coro: support listed coroutines
QEMU's coroutines are really hard to find, but this is required
so that patching won't ruin the code being executed.

Tracking coroutines is much easier from inside the QEMU, but requires
some tricks such as exporting head of the list of the running
coroutines and information about coroutine's internals.

Support use of this type of info from our side.

Change-Id: I5806b0a927ec2ac2ca525eff8fe53740a6acf589
Signed-off-by: Pavel Boldin <pboldin@cloudlinux.com>